Fix filename case collisions for RealEsrganProcessor
- Renamed RealEsrganProcessor.h/cpp to lowercase realesrganprocessor.h/cpp to match .pro file and resolve Git case collision warnings.
- Updated #include directives to use lowercase filenames.
- Corrected '// file:' comments in header and source file.
